requirements.txt
pandas
sqlalchemy
psycopg2-binary
requests
beautifulsoup4
warcio
rapidfuzz
prefect
dbt-core
dbt-postgres
openai
python-dotenv
